Removed most of the current_user caching attempt.
Was a poor attempt and didn't know what I was doing.

Have also improved the role detection code to do less DB hits.

Eventually I hope to store the user ID in the session, and use that as
the key to access a cached user + roles data. But lets worry about that
when it seems necessary. I find myself guilty of overoptimising!