Refactored follow / unfollow buttons
- Removed duplicate view code from JS response (now calls the
follow_button method)
- Simplified routes so its now just artist_follow and artist_unfollow.

@Jellyfishboy  I'm not sure why but the unfollow button still doesn't
replace the follow button. It must be a JS issue. My bet is that you
don't want to add binds on each click, but rather move that stuff to
application.js and just change class / data attributes in the button so
the correct bind is fired. I don't completely understand whats going on
so I'll leave that for you!